chapter_201/todo_list/src/main/java/ru/job4j/todo/models/Item.java
package ru.job4j.todo.models;

import com.fasterxml.jackson.annotation.JsonFormat;

import java.sql.Timestamp;

/**
 * Item
 *
 * @author Sergey Filippov (serdg1984@yandex.ru).
 * @version 1.0.
 * @since 2019-02-11
 */
public class Item {

    private int id;

    private String desc;

    @JsonFormat(pattern = "dd-MM-yyy HH:mm", timezone = "GMT+3")
    private Timestamp created;

    private boolean done;

    public int getId() {
        return id;
    }

    public void setId(int id) {
        this.id = id;
    }

    public String getDesc() {
        return desc;
    }

    public void setDesc(String desc) {
        this.desc = desc;
    }

    public Timestamp getCreated() {
        return created;
    }

    public void setCreated(Timestamp created) {
        this.created = created;
    }

    public boolean isDone() {
        return done;
    }

    public void setDone(boolean done) {
        this.done = done;
    }

    @Override
    public String toString() {
        return String.format("Item{id=%d, desc=%s, created=%s, done=%s}", id, desc, created, done);
    }
}
